requirements.txt
flask
flask-cors
gunicorn
discord.py==2.3.2
requests
